Supporting buffered files for http bodies

+++ Sources/Protocols/HTTP/Models/HTTPHeader.swift
@@ -121,8 +121,8 @@ extension Dictionary where Key == HTTPHeaderName, Value == String {
     set { self[.contentEncoding] = newValue }
   }
 
-  public var contentLength: Int? {
-    get { return Int(self[.contentLength] ?? "") }
+  public var contentLength: Int64? {
+    get { return Int64(self[.contentLength] ?? "") }
     set { self[.contentLength] = newValue == nil ? nil : "\(newValue!)" }
   }
